# remove logging from router.py