In my experience this is making tests (especially jubjub tests, which are slow) up to 33% faster