@amiller 2019-04-23
There's something suspicious with what I did. First of all, when I ran this test to check that it was adding overhead even when it shouldn't, I changed the log level to `pytest -vs --log-level=INFO`, but that's changing logging in the wrong direction. To exclude the debug message, I should have set it to `--log-level=WARNING`. 

But a second problem... 
a) Even when I change this code to `logging.info`, I still see info messages printed out in the test
b) Even when I change `pytest -vs --log-level=WARNING`, I still see debug messages.
So, I think the real problem is that I'm misundersatnding something about how pytest logging is run. Will update later...