@smkuls 2019-04-23
The default log level for tests is specified in `pytest.ini`. If you don't want debug messages and only want info messages you set the log level to INFO.