@Drake-Eidukas 2019-04-23
I was curious, so I compiled some statistics from my machine:

# Speedup
I was curious, so I compiled some statistics from my machine.
I ran using `docker-compose run --rm honeybadgermpc pytest -v` with the following parallelisms

## Original code with DEBUG test log level

- `-nauto`: 53.37s
- `-n0`: 106.51s

---

## With no logging in router

- `-nauto`: 59.24s
- `-n0`: 92.31s

---

## With INFO test logging level

- `-nauto`: 50.92s
- `-n0`: 96.74s

---

## With WARNING test logging level

- `-nauto`: 51.0s
- `-n0`: 95.47

---

## With WARNING test logging level and no logging in router

- `-nauto`: 50.65s
- `-n0`: 94.42s