@Drake-Eidukas 2019-04-23
I'd be curious what the results of editing pytest.ini would be for you @amiller, as I don't get quite the drastic speedup you do by commenting this out. 

Regardless, we should probably set travis's logging to CRITICAL (if something fails in travis, we're going to re-run it locally, and we can set this to DEBUG if we need to debug something in travis).
Personally, I've actually been using this debug statement for debugging, it's actually quite useful when debugging code hangs for me