@smkuls 2019-04-23
Looks like python's logger is not as smart as `log4j` which actually omits the entire statement when you generate a non debug build.

This seems to be our best bet of solving this:
https://stackoverflow.com/questions/17579775/checking-for-debug-and-some-other-condition-in-python

I wouldn't want to comment the logging out completely because there are times when you need to actually see what messages are being sent/received.