@Drake-Eidukas 2019-04-23
Closing this, as this has been carried out in #249